Various fixes based on feedback in training
Add an `rm K${domain}*` command to remove any existing DNSSEC keys before
generating a new key.  Without this step, if the user runs the
dnssec-keygen command more than once, the KEY variable may be set to an
incorrect value with the wrong key or multiple keys.

Separate the `rndc-confgen` command from the DNSSEC key generation and
briefly explain the purpose of the rndc key.

s/SELinux context/&s/ where appropriate.

Fix the `cat /var/named/dynamic/${domain}.db` command for verifying that
the BIND database was created (the pathname was missing).

Be more explicit in a few references to the broker host or the node host.

Add a missing semicolon to the example dhclient-eth0.conf file, and
emphasise the importance of avoiding typos in this file.

In the verification step for the ActiveMQ console, clarify that one need
use _either_ the `lokkit` command or the SSH port forwarding command, and
not both.

Delete the step on modifying ServerName in the broker's Apache
configuration; this step is not needed with OSE 1.1.

Add a sentence explaining that the broker uses the rsync_id_rsa SSH key
pair for copying gear data from one node host to another for gear
migrations.

Add a brief explanation of what `fixfiles` and `restorecon` do.

Refer to the correct command when describing the step to create the
openshift-origin-dns-bind.conf file (we use `cat`, not `echo`).

Mention that the hairy command for modifying the PAM configuration is
available as a script on the training instances.

Reassure the user that it is safe to ignore error messages from `sysctl`
about unknown keys, as long as those error messages are not related to the
settings we are adding.

Just tell the user to use RubyGems to install the OpenShift Enterprise
client tools on Fedora or RHEL.  Yum repos for Fedora 16 and RHEL have
ancient versions of rhc which don't work.

Fix the heading for Lab 32 to be formatted consistently with the other
labs' headings.

Fix a few grammatical errors.